+++ test/3155-ui_duplicate_point_ctrl_d_test.dart
@@ -8,13 +8,12 @@ import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 import 'package:mapiah/src/controllers/th2_file_edit_selection_controller.dart';
 import 'package:mapiah/src/elements/th_element.dart';
 import 'package:mapiah/src/elements/th2_file.dart';
-import 'package:mapiah/src/generated/i18n/app_localizations.dart';
 import 'package:mapiah/src/generated/i18n/app_localizations_en.dart';
 import 'package:mapiah/src/mp_file_read_write/th_file_writer.dart';
-import 'package:mapiah/src/pages/th2_file_edit_page.dart';
 import 'package:mapiah/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state.dart';
 import 'package:path_provider_platform_interface/path_provider_platform_interface.dart';
 
+import 'th2_file_tabs_page_test_aux.dart';
 import 'th_test_aux.dart';
 
 class _FakePathProviderPlatform extends PathProviderPlatform {
@@ -64,15 +63,7 @@ void main() {
       final TH2File snapshotOriginal = TH2File.fromMap(th2File.toMap());
 
       await tester.pumpWidget(
-        MaterialApp(
-          localizationsDelegates: AppLocalizations.localizationsDelegates,
-          supportedLocales: AppLocalizations.supportedLocales,
-          locale: const Locale('en'),
-          home: TH2FileEditPage(
-            filename: testFilename,
-            th2FileEditController: th2Controller,
-          ),
-        ),
+        buildTH2FileTabsPageTestApp(th2FileEditController: th2Controller),
       );
       await tester.pump();